Refactored code into features : ar, details and furniture
