+++ app/src/main/java/com/example/hsh/homesweethome/details/FurnitureDetails.java
@@ -1,20 +1,19 @@
-package com.example.hsh.homesweethome;
+package com.example.hsh.homesweethome.details;
 
 import android.content.Intent;
-import android.net.Uri;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.v4.view.ViewPager;
 import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
-import android.view.Menu;
-import android.view.MenuItem;
 
+import com.example.hsh.homesweethome.ar.FurnitureARActivity;
 import com.example.hsh.homesweethome.Models.Furniture;
+import com.example.hsh.homesweethome.R;
+import com.example.hsh.homesweethome.furniture.FurnitureFragmentAdapter;
 import com.luseen.spacenavigation.SpaceItem;
 import com.luseen.spacenavigation.SpaceNavigationView;
 import com.luseen.spacenavigation.SpaceOnClickListener;
-import com.squareup.picasso.Picasso;
 
 
 public class FurnitureDetails extends AppCompatActivity{
