+++ app/src/main/java/com/example/hsh/homesweethome/details/FurnitureDetailsMainFragment.java
@@ -1,4 +1,4 @@
-package com.example.hsh.homesweethome;
+package com.example.hsh.homesweethome.details;
 
 import android.annotation.SuppressLint;
 import android.graphics.Rect;
@@ -21,6 +21,7 @@ import android.widget.RatingBar;
 import android.widget.TextView;
 
 import com.example.hsh.homesweethome.Models.Furniture;
+import com.example.hsh.homesweethome.R;
 import com.squareup.picasso.Picasso;
 
 import java.util.ArrayList;
