+++ app/src/main/java/com/example/hsh/homesweethome/furniture/FurnitureCategoryFragment.java
@@ -1,4 +1,4 @@
-package com.example.hsh.homesweethome;
+package com.example.hsh.homesweethome.furniture;
 
 import android.app.Dialog;
 import android.os.Bundle;
@@ -17,6 +17,7 @@ import android.widget.SeekBar;
 import android.widget.TextView;
 
 import com.example.hsh.homesweethome.Models.CategoryFurniture;
+import com.example.hsh.homesweethome.R;
 
 import java.util.ArrayList;
