+++ app/src/main/java/com/example/hsh/homesweethome/furniture/FurnitureFragmentAdapter.java
@@ -1,12 +1,14 @@
-package com.example.hsh.homesweethome;
+package com.example.hsh.homesweethome.furniture;
 
 import android.content.Context;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
 import android.support.v4.app.FragmentStatePagerAdapter;
-import android.util.Log;
 import android.view.ViewGroup;
 
+import com.example.hsh.homesweethome.details.FurnitureDetailsCartFragment;
+import com.example.hsh.homesweethome.details.FurnitureDetailsMainFragment;
+import com.example.hsh.homesweethome.details.FurnitureDetailsReviewFragment;
 import com.example.hsh.homesweethome.Models.Furniture;
 
 public class FurnitureFragmentAdapter extends FragmentStatePagerAdapter {
