+++ app/src/main/java/com/example/hsh/homesweethome/furniture/RecyclerViewAdapterLocations.java
@@ -1,4 +1,4 @@
-package com.example.hsh.homesweethome;
+package com.example.hsh.homesweethome.furniture;
 
 import android.content.Context;
 import android.support.annotation.NonNull;
@@ -7,6 +7,8 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
+
+import com.example.hsh.homesweethome.R;
 
 import java.util.ArrayList;
