+++ app/src/main/java/com/example/hsh/homesweethome/furniture/RecyclerViewAdapterMain.java
@@ -1,4 +1,4 @@
-package com.example.hsh.homesweethome;
+package com.example.hsh.homesweethome.furniture;
 
 
 import android.content.Context;
@@ -20,6 +20,7 @@ import android.widget.TextView;
 
 import com.example.hsh.homesweethome.Models.CategoryFurniture;
 import com.example.hsh.homesweethome.Models.Furniture;
+import com.example.hsh.homesweethome.R;
 
 import java.util.ArrayList;
 import java.util.Arrays;
