+++ app/src/main/res/layout/content_main.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context=".MainActivity"
+    tools:context=".furniture.MainActivity"
     tools:showIn="@layout/app_bar_main">
 
     <android.support.v7.widget.CardView
